requirements.txt
asgiref==3.4.1
Django==3.1.7
django-crispy-forms==1.12.0
django-crum==0.7.9
django-js-asset==1.2.2
django-mptt==0.13.2
django-widget-tweaks==1.4.8
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
